fix deleting self
