+++ source/modules/argument_parsing.py
@@ -9,15 +9,15 @@ from modules._platform import is_frozen, show_windows_help
 
 def error(parser: ArgumentParser, msg: str):
     if is_frozen() and sys.platform == "win32":
-       	from PySide6.QtWidgets import QApplication
-        from windows.popup_window import PopupWindow, PopupIcon
+        from PySide6.QtWidgets import QApplication
+        from windows.popup_window import PopupIcon, PopupWindow
 
         app = QApplication([])
         PopupWindow(
             title="Error",
             message="An error occurred during parsing arguments:<br>" + parser.format_usage() + "<br>" + msg,
             icon=PopupIcon.WARNING,
-            button="Quit",
+            buttons=["Quit"],
             app=app,
         ).show()
         sys.exit(app.exec())
